\name{Rvk-package}
\alias{Rvk-package}
\alias{Rvk}
\docType{package}
\title{
\packageTitle{Rvk}
}
\description{
\packageDescription{Rvk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rvk}
\packageIndices{Rvk}
The package uses

}
\author{
\packageAuthor{Rvk}

Maintainer: \packageMaintainer{Rvk}
}


\references{
VK API: Open API (\url{https://vk.com/dev/openapi})
}


~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }

\examples{
~~ simple examples of the most important functions ~~
}
