% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{rvk_pal}
\alias{rvk_pal}
\title{rvk_pal
Litapaletta Gagnaþjónustu Reykjavíkurborgar. Hún er ein og hefur 9 gildi.}
\usage{
rvk_pal()
}
\description{
rvk_pal
Litapaletta Gagnaþjónustu Reykjavíkurborgar. Hún er ein og hefur 9 gildi.
}
