% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvk_palettur.R
\docType{data}
\name{rvk_palettur}
\alias{rvk_palettur}
\title{Litapalettur fyrir þemu}
\format{Sem stendur er þetta \code{tibble} object.}
\usage{
rvk_palettur
}
\description{
\code{rvk} pakkinn inniheldur ýmis gildi fyrir litapalettur.
}
\keyword{datasets}
