% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_rvk}
\alias{scale_color_rvk}
\title{scale_color_rvk
Litaskali Gagnaþjónustu Reykjavíkurborgar}
\usage{
scale_color_rvk(...)
}
\description{
scale_color_rvk
Litaskali Gagnaþjónustu Reykjavíkurborgar
}
