% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_rvk}
\alias{scale_fill_rvk}
\title{scale_fill_rvk
Litaskali Gagnaþjónustu Reykjavíkurborgar}
\usage{
scale_fill_rvk(...)
}
\description{
scale_fill_rvk
Litaskali Gagnaþjónustu Reykjavíkurborgar
}
