% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rvk.R, R/themes.R
\name{theme_rvk}
\alias{theme_rvk}
\title{Fyrra ggplot2 þema Gagnaþjónustu Reykjavíkurborgar}
\usage{
theme_rvk(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  ...
)

theme_rvk(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  ...
)
}
\arguments{
\item{base_size}{grunnstærð leturs}

\item{...}{inntök sem færð eru áfram í \code{theme_minimal}}
}
\value{



}
\description{
Þetta er fyrsta uppkast að sérstöku þema fyrir Gagnaþjónustu Reykjavíkurborgar.
Sem stendur er þetta að bara umritað theme_roboto úr silgelib pakkanum hennar Juliu Silge.

Þetta er fyrsta uppkast að sérstöku þema fyrir Gagnaþjónustu Reykjavíkurborgar.
Sem stendur er þetta að bara umritað theme_roboto úr silgelib pakkanum hennar Juliu Silge.
}
\details{
Roboto Condensed og Roboto Bold letrin eru Google letur;
hægt er að hala þeim niður á \url{https://fonts.google.com/specimen/Roboto+Condensed}
og \url{https://fonts.google.com/specimen/Roboto}. Þessar leturgerðir verða
að vera uppsettar á tölvunni þinni til að þú getir notast við þemað.

Roboto Condensed og Roboto Bold letrin eru Google letur;
hægt er að hala þeim niður á \url{https://fonts.google.com/specimen/Roboto+Condensed}
og \url{https://fonts.google.com/specimen/Roboto}. Þessar leturgerðir verða
að vera uppsettar á tölvunni þinni til að þú getir notast við þemað.
}
\examples{
ggplot(mtcars, aes(x = drat, y = mpg ))
ggplot(mtcars, aes(x = drat, y = mpg ))
}
