% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_rvk2}
\alias{theme_rvk2}
\title{Síðara ggplot2 þema Gagnaþjónustu Reykjavíkurborgar}
\usage{
theme_rvk2(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  ...
)
}
\arguments{
\item{base_size}{grunnstærð leturs}

\item{...}{inntök sem færð eru áfram í \code{theme_minimal}}
}
\description{
Síðara ggplot2 þema Gagnaþjónustu Reykjavíkurborgar
}
\details{
IBM Plex leturgerðirnar eru open source og hægt er að nálgast þær á
\url{https://github.com/IBM/plex}. Þessar leturgerðir verða
að vera uppsettar á tölvunni þinni til að þú getir notast við þemað.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?") +
    theme_rvk2()

ggplot(diamonds, aes(carat, price, color = clarity)) +
    geom_point() +
    facet_wrap(~cut) +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?") +
         theme_rvk2()

}

}
