\name{Rvk-package}
\alias{Rvk-package}
\alias{Rvk}
\docType{package}
\title{
\packageTitle{Rvk}
}
\description{
\packageDescription{Rvk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rvk}
\packageIndices{Rvk}
The package uses

}
\author{
\packageAuthor{Rvk}

Maintainer: \packageMaintainer{Rvk}
}


\references{
VK API: Open API (\url{https://vk.com/dev/openapi})
}


\keyword{ package }


